add `stream-utils` sub-import with the `createInjectionTransformStream` and `pipeReaderToResponse` helpers
