+++ packages/client-react-streaming/package.json
@@ -76,12 +76,31 @@
         "node": "./dist/manual-transport.ssr.js"
       }
     },
+    "./stream-utils": {
+      "require": {
+        "types": "./dist/stream-utils.node.d.cts",
+        "edge-light": "./dist/stream-utils.node.cjs",
+        "react-server": "./dist/empty.cjs",
+        "browser": "./dist/empty.cjs",
+        "node": "./dist/stream-utils.node.cjs"
+      },
+      "import": {
+        "types": "./dist/stream-utils.node.d.ts",
+        "edge-light": "./dist/stream-utils.node.js",
+        "react-server": "./dist/empty.js",
+        "browser": "./dist/empty.js",
+        "node": "./dist/stream-utils.node.js"
+      }
+    },
     "./package.json": "./package.json"
   },
   "typesVersions": {
     "*": {
       "manual-transport": [
         "./dist/manual-transport.ssr.d.ts"
+      ],
+      "stream-utils": [
+        "./dist/stream-utils.node.d.ts"
       ]
     }
   },
