+++ packages/client-react-streaming/src/stream-utils/index.ts
@@ -1,0 +1,2 @@
+export { createInjectionTransformStream } from "./createInjectionTransformStream.js";
+export { pipeReaderToResponse } from "./pipeReaderToResponse.js";
